testing/web-platform/tests/battery-status/battery-discharging-manual.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Battery Status Test: BatteryManager - discharging</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="http://www.w3.org/TR/battery-status/">
<meta name="flags" content="interact">
<meta name="assert" content="All of the navigator.battery attributes exist and are set to their correct values when battery is discharging">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
  <h2>Preconditions</h2>
  <ol>
    <li>
      The device is unplugged from the charger before this test is run.
    </li>
    <li>
      The battery must neither be empty or full, nor reach empty or full capacity during the test.
    </li>
  </ol>
</body>
<script>

setup({ explicit_timeout: true });

async_test(function (t) {
  navigator.getBattery().then(function (battery) {
      t.step(function () {
        assert_false(battery.charging, 'The BatteryManager.charging must be set to false');
        assert_equals(battery.chargingTime, Infinity, 'The BatteryManager.chargingTime');
        assert_less_than(battery.dischargingTime, Infinity, 'The BatteryManager.dischargingTime');
        assert_greater_than(battery.level, 0, 'The BatteryManager.level');
        assert_less_than(battery.level, 1.0, 'The BatteryManager.level');

        var battery_level = battery.level;
        battery.onlevelchange = t.step_func(function (e) {
          assert_less_than(e.level, battery_level, 'The value of the level attribute must decrease');
          t.done();
        });
      });
    }, function (error) {
      t.step(function () {
        assert_unreached(error.message);
      });
      t.done();
    }
  )
}, document.title);

</script>
